dataplane: derive NAT counter_id from rule identity for cross-compile stability
The per-rule NAT translation-hit counter_id was assigned sequentially
(nextNATCounterID reset to 1 each compile) while the Rust helper keeps a
cumulative numeric-keyed counter store (NatCounterStore::reconcile_ids
retains stored ids, it does not reset them). A config reorder or the
removal/re-add of a rule could therefore reuse a numeric counter_id for a
DIFFERENT rule, and the new rule inherited the old rule's cumulative
count -- cross-rule mis-attribution in `show security nat ... rule`
Translation hits. This is display-only (never affects forwarding or
translation correctness) and resets on helper restart, but it is wrong.
Follow-up from #2249 (#2218); same class as the closed #1760.

Fix (issue #2255 option a): make counter_id a function of the rule's
IDENTITY, not its config position. assignNATCounterID now derives the id
as a 32-bit FNV-1a hash of the already-namespaced NATCounterKey
(natType/ruleset/rule, introduced in #2249) via natCounterIDForKey,
instead of the sequential nextNATCounterID. A rule keeps the same id
across a reorder or the removal of an unrelated rule, so the cumulative
helper store stays correctly attributed by construction -- reconcile_ids
retains an id that always refers to the same rule it did before.

The counter_id chain is widened u16 -> u32 (CompileResult/ApplyResult
maps, the SNAT/DNAT/static snapshot fields, the Rust NatCounterStore key,
snapshot_active_nat_counter_ids, and the status report) so the wide hash
space makes a distinct-key collision negligible (~256^2/2^32 ~ 8e-6 at
the 256-rule cap). The rare in-compile collision is resolved
deterministically (re-hash with a "#N" suffix) so every rule still gets a
unique id, reproduced identically on every compile. The exhaustion
contract is preserved: past MaxNATRuleCounters distinct counters a rule
falls back to counter 0.

No wire change. The control protocol is serde JSON and a JSON number is
integer-width agnostic, so widening the field needs no protocol_wire_v1
regen -- wire_invariant_default_specimens passes unchanged.

ReadNATRuleCounter now keys the sparse u32 offset map directly and no
longer indexes the legacy 256-entry nat_rule_counters BPF array: a hash
id >= MaxNATRuleCounters would fail that bounded Lookup, and the array
only ever held zeros at runtime (the Rust forwarder reports hits via the
offset map; the #1476 eBPF retirement dropped the XDP increment), so
dropping the Lookup changes no observable value. The legacy
snat_value.counter_id BPF struct field is now vestigial in the userspace
runtime (written through a no-op DataPlane) and left untouched.

Tests (fail-on-revert): TestAssignNATCounterIDStableAcrossReorder asserts
each rule's id is unchanged when rules are reordered (RED under a
simulated sequential scheme, which swaps the ids); StableAcrossRemoval
asserts the survivor keeps its id when an unrelated earlier rule is
removed; Deterministic asserts the same key derives the same id across
independent compiles; UniqueAtScale asserts every id is unique at the cap;
ExhaustionFallsBackToZero preserves the cap fallback. Rust
nat_counter_store_stable_ids_survive_reorder_and_removal proves a retained
id keeps its count and a removed id is dropped so a re-added rule cannot
inherit it. Updated the existing collision/apply/manager/per-uplink tests
for the u32 type. Updated the NAT per-rule-counter contract in
docs/userspace-dataplane-gaps.md.

Refs #2255.

+++ pkg/dataplane/userspace/manager_test.go
@@ -5519,7 +5519,7 @@ func TestClearNATRuleCountersSendsHelperIPCAndIsDurable(t *testing.T) {
 	}
 	defer ln.Close()
 
-	const counterID = uint16(1)
+	const counterID = uint32(1)
 	const cumPackets = uint64(42)
 	const cumBytes = uint64(4200)
